Add slope and restriction parameters for movement

+++ Expeditionary/Model/Mapping/Tile.cs
@@ -4,8 +4,8 @@ namespace Expeditionary.Model.Mapping
 {
     public record class Tile
     {
-        public float Elevation { get; set; }
-        public float Slope { get; set; }
+        public int Elevation { get; set; }
+        public int Slope { get; set; }
         public float Heat { get; set; }
         public float Moisture { get; set; }
         public Terrain Terrain { get; set; } = new();
